Refactor definitions
