+++ src/definition/ui/UIActionButtonContext.ts
@@ -1,0 +1,6 @@
+export enum UIActionButtonContext {
+    MESSAGE_ACTION = 'messageAction',
+    ROOM_ACTION = 'roomAction',
+    MESSAGE_BOX_ACTION = 'messageBoxAction',
+    SIDEBAR_ITEM = 'sidebarItem'
+}
