+++ src/server/managers/UIActionButtonManager.ts
@@ -1,29 +1,4 @@
-import { RoomType } from '../../definition/rooms/RoomType';
-
-export enum ActionButtonContext {
-  MESSAGE_ACTION = 'messageAction',
-  ROOM_ACTION = 'roomAction',
-  MESSAGE_BOX_ACTION = 'messageBoxAction',
-  SIDEBAR_ITEM = 'sidebarItem',
-}
-
-export interface IUIActionButtonDescriptor {
-  actionId: string;
-  nameI18n: string;
-  hintI18n: string;
-  context: ActionButtonContext;
-  // Suggestions for possible icons?
-  icon?: string;
-  when?: {
-    roomTypes?: Array<RoomType>;
-    // How do we provide suggestions for permissions?
-    hasOnePermission?: Array<string>;
-    hasAllPermissions?: Array<string>;
-    // How do we provide suggestions for roles?
-    hasOneRole?: Array<string>;
-    hasAllRoles?: Array<string>;
-  };
-}
+import { IUIActionButtonDescriptor } from '../../definition/ui';
 
 export class UIActionButtonManager {
   private registeredActionButtons = new Map<string, Map<string, IUIActionButtonDescriptor>>();
@@ -38,5 +13,13 @@ export class UIActionButtonManager {
 
   public clearAppActionButtons(appId: string) {
     this.registeredActionButtons.set(appId, new Map());
+  }
+
+  public getAppActionButtons(appId: string) {
+      return this.registeredActionButtons.get(appId);
+  }
+
+  public getAllActionButtons() {
+      return this.registeredActionButtons;
   }
 }
